Code for "Analyzing how habitat degradation drives extinction dynamics using physiologically-structured population models"

+++ src/cuda/species/inds_stochastic_migratory.cu
@@ -1,5 +1,5 @@
-#include "inds_stochastic_migratory.h"
-#include "file_checker.h"
+#include <species/inds_stochastic_migratory.h>
+#include <util/file_checker.h>
 
 inds_stochastic_migratory::inds_stochastic_migratory(int size_val, int maxsize_val, int seed_val, int ndemes,  int species_ID_val) : inds_stochastic(size_val, maxsize_val, seed_val, ndemes, species_ID_val)
 	{
@@ -9,7 +9,7 @@ inds_stochastic_migratory::inds_stochastic_migratory(int size_val, int maxsize_v
 *
 */ 
 	// specify the migration rate from subpopulation i to subpoulation j
-	thrust::device_vector<float> migration_probability(ndemes*ndemes);
+	thrust::device_vector<double> migration_probability(ndemes*ndemes);
 
 	// for convenience, suppose that the migration matrix is read in from a file Migration_Probabilities.txt
 	FILE *Migration_Probabilities_File; 
@@ -27,8 +27,8 @@ inds_stochastic_migratory::inds_stochastic_migratory(int size_val, int maxsize_v
 		
 	for (int i=0; i < ndemes*ndemes; i++)
 		{	
-		float test;
-		fscanf(Migration_Probabilities_File, "%f\n", &test);
+		double test;
+		fscanf(Migration_Probabilities_File, "%lf\n", &test);
 		migration_probability[i] = test;
 		}
